Finish some user experience mods, and building .exe version
Fixing a few more bugs, particularly modes for Econs with no trade for
1,2,3 or more years, and Econs with trades for 1,2,3 or more years
These counts help look at the positive life effects of trading.
I also used launch4j application to create a windows .exe version
of StarTrader containing a JDK so no preloaded version of Java is
needed.  The JDK included is the Oracle free version of JDK_19.0.2
and the use of that JDK is subject to the "No-Fee Terms and Conditions"