move to version 19.58 start class Strategy
add routines in E. to set the character in the key according to the number for the value.  Also a routine in E to get the number associated with the 
character in the ky.
Start defining the class Strategy which will contain methods to select a strategy with the most joy, suggesting settings to help increase your settings and score.  Useful for setting the ship strategy at the start of the year, and the tradeFrac for smart clans