Continue debug AI setCntAr version 19.63 more fixes
I fixed the result of the tests for value, to instead of the first entry less than the input value, add one for the last entry greater than input value except for the largest value which cannot exceed the size of the array.  I have the reading of the mapfile working, so not entries are kept between runs.  The mapfile exists in the base folder which also has the keep file.  I do not commit the mapfile or keep file